refactor(style): :recycle: Added dependency inversion principle to PlayerService
