+++ griffindor-backend/src/main/java/com/devathon/griffindor_backend/services/PlayerService.java
@@ -1,69 +1,24 @@
 package com.devathon.griffindor_backend.services;
 
-import com.devathon.griffindor_backend.config.WebSocketRoutes;
 import com.devathon.griffindor_backend.enums.PlayerSessionState;
-import com.devathon.griffindor_backend.events.PlayerCountUpdateEvent;
 import com.devathon.griffindor_backend.models.Player;
 import java.util.Collection;
-import java.util.concurrent.ConcurrentHashMap;
-import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.stereotype.Service;
 
-@Service
-public class PlayerService {
+public interface PlayerService {
 
-    private final ApplicationEventPublisher eventPublisher;
-    private final ConcurrentHashMap<String, Player> players = new ConcurrentHashMap<>();
+    void addPlayer(String sessionId, String name, String house, PlayerSessionState sessionState);
 
-    public PlayerService(ApplicationEventPublisher eventPublisher) {
-        this.eventPublisher = eventPublisher;
-    }
+    void removePlayer(String sessionId);
 
-    public void addPlayer(String sessionId, String name, String house, PlayerSessionState sessionState) {
-        Player player = new Player(sessionId, name, house, sessionState);
-        players.put(sessionId, player);
-        publishPlayerCountUpdate();
-    }
+    void updatePlayerInfo(String sessionId, String name, String house);
 
-    public void removePlayer(String sessionId) {
-        players.remove(sessionId);
-        publishPlayerCountUpdate();
-    }
+    void updatePlayerSessionState(String sessionId, PlayerSessionState state);
 
-    public void updatePlayerInfo(String sessionId, String name, String house) {
-        Player player = players.get(sessionId);
-        if (player != null) {
-            player.setName(name);
-            player.setHouse(house);
-        }
-    }
+    int getActivePlayerCount();
 
-    public void updatePlayerSessionState(String sessionId, PlayerSessionState state) {
-        Player player = players.get(sessionId);
-        if (player != null) {
-            player.setSessionState(state);
-        }
-    }
+    void publishPlayerCountUpdate();
 
-    public int getActivePlayerCount() {
-        return players.size();
-    }
+    void printAllPlayers();
 
-    public void publishPlayerCountUpdate() {
-        eventPublisher.publishEvent(new PlayerCountUpdateEvent(players.size(), WebSocketRoutes.TOPIC_NUM_PLAYERS));
-    }
-
-    public void printAllPlayers() {
-        System.out.println("📑 Players: ");
-        for (Player player : players.values()) {
-            System.out.println("Player ID: " + player.getSessionId() + ", Name: "
-                    + player.getName() + ", House: "
-                    + player.getHouse() + ", SessionStatus: "
-                    + player.getSessionStatus());
-        }
-    }
-
-    public Collection<Player> getAllPlayers() {
-        return players.values();
-    }
+    Collection<Player> getAllPlayers();
 }
